+++ Geoscape/RecruitInfo/changelog.md
@@ -1,10 +1,10 @@
 Changelog of Recruit Info, a Phoenix Point Mod by Sheepy
 
-# Version 2
+# Version 2, 2020-04-17
 
+* New: Show recruit equipments (name and simple description only, no stats).
 * New: Show mutations and implants.
-* New: Show recruit equipments (name and simple description only).
-* New: Config to change font size and what info to display.
+* New: Config to what info to displan and font size/style.
 * Fix: PPML 0.1 compatibility
 
 # Version 1, 2019-12-27